Broke up element.h into element.h and elibase.h.
elibase.h contains all the code common to the old and new ELI.
This is in preparation for adding the new ELI.
